reorganized platforms info
