+++ app/components/steps/ExportInstructions.tsx
@@ -1,14 +1,12 @@
 import { PlatformInstructions } from "@app/components/PlatformInstructions";
-import { Platform } from "@pipeline/Types";
-
-import { Platforms } from "@assets/Platforms";
+import { Platform, PlatformsInfo } from "@pipeline/Platforms";
 
 interface Props {
     platform?: Platform;
 }
 
 export const ExportInstructions = ({ platform }: Props) => {
-    const info = platform ? Platforms[platform] : undefined;
+    const info = platform ? PlatformsInfo[platform] : undefined;
 
     return (
         <div className="ExportInstructions">
@@ -21,7 +19,7 @@ export const ExportInstructions = ({ platform }: Props) => {
                     color: `hsl(${info?.color[0]}, ${info?.color[1]}%, ${info?.color[2]}%)`,
                 }}
             >
-                {info?.title}
+                {info?.name}
             </span>
             :
             <br />
